Fix CMakeLists.txt to not exclude everything with 'build' in pathname
Summary:
The cmake globbing then excluding strategy implies that if the directory
in which we are compiling has a folder named `build` on its path like `/home/alice/project/build/folly/...`, everything breaks.
Note that it is the same for `tools` and `test`, I also fixed the tools one, but the test once has many files.... Using a file list would be the real solution.
Closes https://github.com/facebook/folly/pull/815

Reviewed By: simpkins

Differential Revision: D7624035

Pulled By: yfeldblum

fbshipit-source-id: a3f1d2f82e55e070d13acc5c394d8cfdfabeb7ed