cmake: fix error in exception_tracer/CMakeLists.txt
Summary:
D7479058 changed the exception_tracer code to export their targets in the main
folly-targets.cmake file, but I forgot to remove the `install()` command
installing the old `folly-exception-tracer-targets.cmake` file.

This would cause `cmake configure` to exit with an error if it detected it
could built the exception_tracer code, but it would still generate all of the
necessary Makefiles correctly, so subsequent `make` commands would still work.

Reviewed By: yfeldblum

Differential Revision: D7626710

fbshipit-source-id: f015032c3c42cf2911fcaf9b355fc97616ff20e1