Add errmsg callback to udp
Summary:
Adds the ability for UDP sockets to optionally
listen to errors that happen on the socket,
for example ICMP errors.

In this iteration, we explicitly do not register for
read events if an error callback is set. So even
if you set an error callback to get error events
you must register a read callback.

Reviewed By: yfeldblum

Differential Revision: D7632877

fbshipit-source-id: 43c922d0145e1da97b993f5bf4058c20d3469deb