Implement partial loadUnaligned
Summary:
Neither GCC or Clang are able to optimize a `memcpy` where
the length is known to be smaller than a word: a library call is
generated. This is a simple implementation of a subword
`loadUnaligned` to be used to read the tail of a buffer when reading
one word at a time.

Reviewed By: luciang

Differential Revision: D7631978

fbshipit-source-id: d5719580e6bef82aa8e3fbdce8f873b569527a3a