Significantly improve merge performance by leveraging inplace_merge
Summary:
The sorting approach used in TDigest is inefficient because it is not leveraging the fact that there are k sorted subarrays. This diff leverages inplace_merge to improve performance.

The cost of merging k digests of size m used to be O(km * log(km)). It is now O(km * log(k)).

Reviewed By: anakryiko

Differential Revision: D7690143

fbshipit-source-id: 1307db153b3cae0bb952d4b872aede8c40ce292c