fix CMake build to link to libunwind correctly again
Summary:
D7642870 accidentally broke linking of folly when libunwind is available.
It change the code to use the contents of `${FOLLY_LINK_LIBRARIES}` in
folly-deps.cmake, but libunwind was added to `${FOLLY_LINK_LIBRARIES}` after
folly-deps.cmake was included, in FollyConfigChecks.cmake.

This moves the check for libunwind to folly-deps.cmake to fix this issue.  All
of the other checks for external libraries are already in folly-deps.cmake.

Reviewed By: yfeldblum

Differential Revision: D7764762

fbshipit-source-id: 922459a7eefcbd92d3d77c02e9c215bb1a5467cc