BufferedStat
Summary:
BufferedStat hooks up the DigestBuilder and a Digest or a SlidingWindow

Every windowDuration time, the DigestBuilder builds a new digest that will be merged into the central data structure (either a digest or a sliding window of digests).

This diff removes the shared buffer from DigestBuilder, as the concurrency model was getting overly complex with it.

Reviewed By: simpkins

Differential Revision: D7629082

fbshipit-source-id: 54027363d37c0fdced287550eea5b12caf6b8138