add connect method to AsyncUDPSocket
Summary:
Calling connect on the UDP socket can avoid route lookups
on the write path. This would be an optimization that would be
useful for clients.

There are several caveats of using connect() which we discovered
during testing. These are documented and unit tests have been written
for each of them.

Reviewed By: yfeldblum

Differential Revision: D7802389

fbshipit-source-id: 09b71373a3a95c5dab73ee9345db0dbbf66d4ec5