Fix handling time counters
Summary:
Currently we'll are getting incorrect Idle&Busy time counters for loops when there were no LoopCallbacks. This is because we don't properly refresh the `startWork_` and use the one from the last loop that had LoopCallbacks.

This diff fixes that by properly refreshing `startWork_` even if there're no LoopCallbacks.

Reviewed By: yfeldblum

Differential Revision: D7811127

fbshipit-source-id: 94ded08963b1bddc7b7859484f076d4c8a1e897f