FutureBase::assign() consumes its arg, so pass that arg via rvalue-ref
Summary:
future/semifuture move-assignment operators call this->assign(). The latter's arg was passed as an lvalue-ref. Result: now that arg is an rvalue-ref.

Note: I adjusted the assign() calling convention to avoid a forwarding ref. It would have been okay but it's easier to avoid a forwarding-ref than to explain why it happens to always be an rvalue-ref.

Reviewed By: yfeldblum

Differential Revision: D7831539

fbshipit-source-id: cf1c66cd498f768905ac3d21e39f693d7a652b39