destructuring for F14*Set::emplace
Summary:
This diff optimizes the operation of F14 set emplace when
the argument list happens to be a single reference to a key_type.
An intermediate key_type value is removed in that case, which avoids a
move in all cases and avoids a copy when the emplace finds that the key
is already present in the set.

Reviewed By: yfeldblum

Differential Revision: D7845720

fbshipit-source-id: 16a639f53993d9843ffe265edd58c0d74de0c1f7