add valid() to Future, SemiFuture, Promise
Summary:
Lets callers determine whether a Future/SemiFuture/Promise has a shared-state.
`x.valid()` (where `x` is a Future, SemiFuture, or Promise) is is true if/only if the object has a shared state.

The `valid()` method will also be key in a near-term diff that will use it to clarify the contracts/rules.

Note: the concept of validity already shows up in the doc-blocks; this gives callers a conceptual oppty to check it.

Reviewed By: yfeldblum

Differential Revision: D7839038

fbshipit-source-id: f8390b3d1a8c3d5d89356ae48937e03c424d1174