only use -Wno-inconsistent-missing-override if supported
Summary:
Check if the compiler supports -Winconsistent-missing-override before trying to
disable this warning.

It looks like gcc wouldn't actually fail the build if
`-Wno-inconsistent-missing-override` was specified but was not supported.
However, if the build failed due to other compile errors the error messages
about this flag would make it confusing to identify the actual build failure
reason.

Closes issue #836.

Reviewed By: yfeldblum

Differential Revision: D7920577

fbshipit-source-id: 322444a9df3f4f36bb5420f0720c1e433c2856b7