Fix for failing travis CI tests
Summary:
Running the xlog test binary stand-alone revealed a bug that
was otherwise masked when running with buck test. The bug was that
2 tests ended up depending on the log-level associated with a default
category and would thus fail since the tests did not clean up
the log-level setting associated with those categories.
Attached diff fixes the issue by restoring the category's log-level in
a scope exit block.

Reviewed By: simpkins

Differential Revision: D7922086

fbshipit-source-id: 9d8ed917633ae214dc81075fbd38275160e64376