Properly align small_vector.
Summary:
Previously the alignment of small_vector did not take into account the alignment of the value_type. Since the value_type is stored directly internally, this is problematic.

This change does increase the sizeof small_vector in a couple cases, but I don't think the previous size was legitimate as sizeof(small_vector<value_type>) was not an even multiple of alignof(value_type).

Reviewed By: yfeldblum

Differential Revision: D7969483

fbshipit-source-id: 968c78bf03d7c0ddfe1338754f12f97be8d0f771