F14: add an API for enumerating dynamic allocation classes
Summary:
Internal memory fragmentation is responsible for a non-trivial
amount of the memory footprint of a dynamically-allocated hash table.
This diff adds an API visitAllocationClasses that allows the caller to
count exactly how many dynamic allocations were performed and of what
size, so that they can accurately account for overheads not included in
getAllocatedMemorySize's simple sum of the underlying malloc requests.

This diff also exposes F14Table's reset() via F14 maps and sets, which
was an unintentional omission.

Reviewed By: WillerZ

Differential Revision: D7894385

fbshipit-source-id: 2d7cadd3806e3f38a9e38cb94c90492a9965ff47