Workaround MSVC 15.7 bug
Summary:
MSVC 15.7 (VS 2017 Update 7) was released a couple days ago, and includes a bug that's breaking Folly. Temporarily work around it.
A point release should be out soon(ish) to resolve the issue: https://developercommunity.visualstudio.com/content/problem/246602/157-fails-with-folly.html
There is still another issue in 15.7 that needs to be worked around, which will happen in a separate diff.

Reviewed By: yfeldblum

Differential Revision: D7937972

fbshipit-source-id: 4a2bc4ae0ea7165c45d100ad0a29409b60a0287a