Call onSet and onUnset when adding and and deleting RequestData
Summary: Currently, when we add or delete a `RequestData` instance to `RequestContext` for the first time, we do not call the `RequestData`'s `onSet` or `onUnset` callback. This means if the `RequestData` is doing tracking-related work, it may miss the beginning or the request until the first switch of `RequestContext` has happend, and in the end as well. This Diff fixes the issue by calling `onSet` and `onUnset` when adding or deleting.

Reviewed By: anakryiko

Differential Revision: D7978623

fbshipit-source-id: a524286371f890598bc773e957ef609e90086cea