hazptr: Fix cleanup to wait for the concurrent reclamation of objects
Summary:
This fix aims to prevent cases (pointed out by David Goldblatt) where cleanup returns while a concurrent slower bulk_reclaim has not completely reclaimed objects that are supposed to be reclaimed before the return from reclaim.

This fix uses an atomic count of the number of ongoing  bulk_reclaim-s. Cleanup returns only if it observes that that count is zero.

Reviewed By: djwatson

Differential Revision: D7586325

fbshipit-source-id: 341af641d20a7a07206c7238ea0356a9eabdbab5