Avoid errnoeous deprecation warning in F14BasicSet
Summary:
[Folly] Avoid errnoeous deprecation warning in `F14BasicSet`.

The deprecation warning is emitted for instantiations like `F14FastSet<std::string>`.

The warning is triggered by looking for move-constructibility of the mapped type for maps or `void` for sets, but `void` is not move-constructible. There are a variety of techniques to solve this, and the smallest change is to use `lift_unit_t` in the right place. `Unit` is move-constructible where `void` is not.

Reviewed By: nbronson

Differential Revision: D8178045

fbshipit-source-id: 1ca06e84b4ed7acd415901afcbe80663513063ea