Fix small_vector on gcc 49
Summary:
In folly::small_vector, we were (mistakenly) relying on a default constructor
for non-trivially constructible types.

The problem got worse on gcc 4.9, because for that compiler version, we set
folly::is_trivially_copyable to be std::is_trivial. This is mostly okay, as
std::is_trivial implies std::is_trivially_copyable. The problem is that
std::is_trivial also implies std::is_default_constructuble, which which was
breaking folly::small_vector

Reviewed By: yfeldblum

Differential Revision: D8346780

fbshipit-source-id: 9c8a9dab21ea2d4fd34f636458324915c30e10a2