Fix libfollylogging link dependency
The folly logging library introduced a new shared library:
libfollylogging. folly::init recently started to depend on on
libfollylogging's initLoggingOrDie(). The problem is that libfolly is
never linked with libfollylogging, leading to the following error
upon application link time:

    /usr/local/lib/libfolly.so: undefined reference to
    folly::initLoggingOrDie(folly::Range<char const*>)

This patch forces libfollylogging to implicitly link with libfolly
through the removal of libfollylogging altogether.