fewer random tests for TDigest
Summary:
The tdigest_test is sometimes timing out on the Travis CI build,
perhaps because it runs the tests sequentially rather than in parallel
as under buck.  This diff reduces the size of the random tests.  At the
reduced size it was slightly exceeding the hard-coded expected accuracy
bounds, so the diff also loosens those.

This is a test-only change.

Reviewed By: yfeldblum

Differential Revision: D8548385

fbshipit-source-id: f78aee6aa22d59b297c2aef757da1dc9e0831052