Backport std::apply and remove folly::applyTuple
Summary:
Add a backport of `std::apply` with a similar implementation that makes it easy
to reason about.  Also implement `applyTuple` in terms of `apply`, `invoke` and
`tuple_cat`

The implementation of `apply` has been taken straight from the standard,
reducing chances of churn in incompatible assumptions (if any).  Note that the
definition and declaration of `apply` has been changed from the way it is
specified in the standard for better SFINAE compatibility, if needed, will
followup with a paper proposing the same change in the standard

Reviewed By: spacedentist

Differential Revision: D8248199

fbshipit-source-id: e3e76193a30b1599eca3fe986a33f61addad7a59