handle allocators with only explicit rebinding constructor
Summary:
D8586283 added an implicit conversion from allocator_type to a
uint8_t-rebound allocator, but allocators might provide only an explicit
rebinding constructor.  This diff makes F14 use the explicit form.
It also adds a binary that when debugged (in a buck build environment)
is useful for evaluating code size.

Reviewed By: shixiao

Differential Revision: D8655289

fbshipit-source-id: c977edb041ce2056f3ce23c52a7f8464a6b9f1bd