fix elision of iter advance in F14 erase with unused return value
Summary:
D8395569 added a prefetch instruction as part of iterator advance
for F14Value and F14Node maps.  This caused the work to advance the
iterator returned by erase to no longer be recognized as dead code by the
GCC optimizer.  This diff separates out advance() and advanceLikelyDead()
methods, avoiding prefetch in the latter case.

Reviewed By: shixiao

Differential Revision: D8656038

fbshipit-source-id: 86cecf40e8622f815d34a98709b90d6ef73dd618