Handle EINVAL errno in MemoryIdler::unmapUnusedStack()
Summary:
If `MemoryIdler` is used with threads that are backed by huge pages, the
`madvise(..., MADV_DONTNEED)` calls will fail with an `EINVAL`. Prior to this
diff, this would lead to an assertion failure. Now, if `EINVAL` is returned, we
warn the user, as `MemoryIdler` may not be able to function as expected.

Reviewed By: yfeldblum

Differential Revision: D8669399

fbshipit-source-id: f97ea7549deffdbf283b3516b9e8feafe18a036c